finder-net/ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.3
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.4.0
supabase==2.3.4
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3
